Module pallet_cairo::types

The Cairo Execution Engine pallet’s runtime custom types. Cairo Execution Engine pallet custom types.

Structs

CairoAssemblyProgram

Cairo assembly program representation. A Cairo assembly program is a program that is compiled from a Sierra program.

CairoAssemblyProgramInput

Cairo assembly program input. This is the input that is passed to the Cairo VM when executing a Cairo assembly program. The input is a vector of vector of bytes. Each individual vector of bytes is a single input and can have a maximum length of MaxCairoAssemblyProgramInputLength. The maximum number of inputs is MaxCairoAssemblyProgramInputNumber.

CairoAssemblyProgramOutput

Cairo assembly program output. This is the output that is returned by the Cairo VM when executing a Cairo assembly program. The output is a vector of vector of bytes. Each individual vector of bytes is a single output and can have a maximum length of MaxCairoAssemblyProgramInputLength. The maximum number of outputs is MaxCairoAssemblyProgramInputNumber.

SierraProgram

Sierra program representation.

Type Definitions

CairoAssemblyProgamId

Identifier of a Cairo assembly program.

SierraProgramId

Identifier of a Sierra program.